# Improve parsing performance for built-in functions