This adds a 1-character lookahead step for parsing builtins so that all of
the alternatives don't need to be tried.

This improves performance by 48% (!) on the example benchmark from #108

Before:

```
time                 251.8 ms   (246.6 ms .. 257.3 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 255.5 ms   (252.9 ms .. 257.2 ms)
std dev              2.562 ms   (1.945 ms .. 3.103 ms)
variance introduced by outliers: 16% (moderately inflated)
```

After:

```
time                 170.4 ms   (169.2 ms .. 172.7 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 173.2 ms   (171.5 ms .. 177.4 ms)
std dev              3.808 ms   (428.7 μs .. 5.659 ms)
variance introduced by outliers: 12% (moderately inflated)
```